feat: dto 사용하도록 변경
